fix: address permission denial review findings
