+++ apps/sva-studio-react/src/routes/admin/instances/-instance-detail-split-sections.test.tsx
@@ -40,7 +40,8 @@ const { mockStudioModuleIamContracts } = vi.hoisted(() => ({
   ],
 }));
 
-vi.mock('../../../lib/plugins', () => ({
+vi.mock('../../../lib/plugins', async (importOriginal) => ({
+  ...(await importOriginal<typeof import('../../../lib/plugins')>()),
   studioModuleIamContracts: mockStudioModuleIamContracts,
 }));
