+++ apps/sva-studio-react/src/routes/admin/modules/-modules-page.test.tsx
@@ -50,7 +50,8 @@ vi.mock('../../../components/ConfirmDialog', () => ({
     ) : null,
 }));
 
-vi.mock('../../../lib/plugins', () => ({
+vi.mock('../../../lib/plugins', async (importOriginal) => ({
+  ...(await importOriginal<typeof import('../../../lib/plugins')>()),
   studioModuleIamContracts: realStudioModuleIamContracts,
 }));
