+++ packages/routing/src/ui-route-access.ts
@@ -1,13 +1,11 @@
 import { redirect } from '@tanstack/react-router';
 import { createPermissionDenialDetails } from '@sva/core';
-import type { PluginDefinition } from '@sva/plugin-sdk';
-
 import type { RouteGuardContext } from './protected.routes.js';
 import { buildInsufficientRoleHref, buildLoginHref } from './protected-route-redirects.js';
-
-type PluginRouteAccessRequirement = NonNullable<
-  PluginDefinition['routes'][number]['accessRequirement']
->;
+import {
+  evaluateRouteAccessRequirement,
+  type PluginRouteAccessRequirement,
+} from './ui-route-access-decision.js';
 
 export type UiRouteAccessRequirements = {
   readonly requiredModuleId?: string;
@@ -28,7 +26,10 @@ export const enforceUiRouteAccessRequirements = async (
     throw redirect({ href: '/?error=auth.insufficientRole' });
   }
 
-  if (requirements.requiredModuleId && !user?.assignedModules?.includes(requirements.requiredModuleId)) {
+  if (
+    requirements.requiredModuleId &&
+    !user?.assignedModules?.includes(requirements.requiredModuleId)
+  ) {
     throw redirect({ href: '/?error=auth.insufficientRole' });
   }
 
@@ -52,14 +53,6 @@ export const enforceUiRouteAccessRequirements = async (
   }
 };
 
-const satisfiesRequiredValues = (
-  required: Readonly<{ mode: 'allOf' | 'anyOf'; values: readonly string[] }>,
-  available: ReadonlySet<string>
-): boolean =>
-  required.values.length > 0 && required.mode === 'allOf'
-    ? required.values.every((value) => available.has(value))
-    : required.values.length > 0 && required.values.some((value) => available.has(value));
-
 export const enforceRouteAccessRequirement = async (
   requirement: PluginRouteAccessRequirement | undefined,
   beforeLoadOptions: {
@@ -75,49 +68,10 @@ export const enforceRouteAccessRequirement = async (
   if (!user) {
     throw redirect({ href: buildLoginHref('/auth/login', beforeLoadOptions.location.href) });
   }
-  if (requirement.kind === 'authenticated') {
-    return;
-  }
-
-  const allowed =
-    requirement.kind === 'platform'
-      ? !user.instanceId && satisfiesRequiredValues(requirement.roles, new Set(user.roles ?? []))
-      : Boolean(user.instanceId) &&
-        user.permissionStatus !== 'degraded' &&
-        (!requirement.moduleId || user.assignedModules?.includes(requirement.moduleId)) &&
-        // Route guards do not receive the resource-scoped authorization evidence needed to
-        // verify a capability. Keep access closed until the capability is evaluated server-side.
-        !requirement.resourceCapability &&
-        satisfiesRequiredValues(requirement.actions, new Set(user.permissionActions ?? []));
-
-  if (!allowed) {
-    const grantedPermissions = new Set(user.permissionActions ?? []);
-    const permissionRequirementSatisfied =
-      requirement.kind === 'tenant'
-        ? satisfiesRequiredValues(requirement.actions, grantedPermissions)
-        : false;
-    const tenantPermissionDenial =
-      requirement.kind === 'tenant' &&
-      requirement.actions.values.length > 0 &&
-      user.permissionActions !== undefined &&
-      user.permissionStatus !== 'degraded' &&
-      (!requirement.moduleId || user.assignedModules?.includes(requirement.moduleId))
-        ? createPermissionDenialDetails({
-            requiredPermissions:
-              requirement.actions.mode === 'allOf' && !permissionRequirementSatisfied
-                ? requirement.actions.values.filter(
-                    (permission) => !grantedPermissions.has(permission)
-                  )
-                : requirement.actions.values,
-            requirementMode: requirement.actions.mode,
-            denialReason:
-              permissionRequirementSatisfied && requirement.resourceCapability
-                ? 'abac_condition_unmet'
-                : 'permission_missing',
-          })
-        : undefined;
+  const decision = evaluateRouteAccessRequirement(requirement, user);
+  if (!decision.allowed) {
     throw redirect({
-      href: buildInsufficientRoleHref('/', 'auth.insufficientRole', tenantPermissionDenial),
+      href: buildInsufficientRoleHref('/', 'auth.insufficientRole', decision.permissionDenial),
     });
   }
 };
